Refactoring mortgage calculation
