+++ apps/calculator-ui-react/src/app/store/mortgage-calculation/mortgageSlice.ts
